+++ src/get-version-by-id.js
@@ -1,10 +1,6 @@
 exports.getVersionById = (versions, id) => {
   if (id != null) {
-    const version = versions.find(version => version.id === id)
-
-    if (version != null) {
-      return version
-    }
+    return versions.find(version => version.id === id)
   }
 
   return [...versions]