+++ src/main.js
@@ -35,7 +35,10 @@ exports.main = async function main() {
       }`)
     }
 
-    core.setOutput('log_entry', version.text)
+    core.setOutput('version', version.id)
+    core.setOutput('date', version.date)
+    core.setOutput('status', version.status)
+    core.setOutput('changes', version.changes)
   }
   catch (error) {
     core.setFailed(error.message)