+++ src/parse-entry.test.js
@@ -33,9 +33,10 @@ test('get readable data from text entry', () => {
 
   expect(output.id).toEqual('v0.0.26')
   expect(output.date).toEqual('2019-02-10')
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('released')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })
 
 test('get readable data from text entry | no title for version', () => {
@@ -47,9 +48,10 @@ test('get readable data from text entry | no title for version', () => {
 
   expect(output.id).toEqual('v0.0.26')
   expect(output.date).toEqual('2019-02-10')
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('released')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })
 
 test('get readable data from text entry | version patern X.X.X', () => {
@@ -61,9 +63,10 @@ test('get readable data from text entry | version patern X.X.X', () => {
 
   expect(output.id).toEqual('0.0.26')
   expect(output.date).toEqual('2019-02-10')
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('released')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })
 
 test('get readable data from text entry | raw version number', () => {
@@ -75,9 +78,10 @@ test('get readable data from text entry | raw version number', () => {
 
   expect(output.id).toEqual('v0.0.26')
   expect(output.date).toEqual('2019-02-10')
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('released')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })
 
 // https://github.com/mindsers/changelog-reader-action/issues/4
@@ -90,9 +94,10 @@ test('get readable data from text entry | unreleased version', () => {
 
   expect(output.id).toEqual('Unreleased')
   expect(output.date).toBeUndefined()
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('unreleased')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })
 
 // https://github.com/mindsers/changelog-reader-action/issues/8
@@ -105,9 +110,10 @@ test('get readable data from text entry | unreleased version', () => {
 
   expect(output.id).toEqual('1.0.0-alpha-a.b-c-somethinglong+build.1-aef.1-its-okay')
   expect(output.date).toEqual('2019-02-10')
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('prereleased')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })
 
 test('get readable data from text entry | unexpected characters between version and date', () => {
@@ -119,9 +125,10 @@ test('get readable data from text entry | unexpected characters between version 
 
   expect(output.id).toEqual('1.0.0-alpha-a.b-c-somethinglong+build.1-aef.1-its-okay')
   expect(output.date).toEqual('2019-02-10')
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('prereleased')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })
 
 // https://github.com/mindsers/changelog-reader-action/issues/10
@@ -134,7 +141,53 @@ test('get readable data from text entry | Fake date', () => {
 
   expect(output.id).toEqual('1.0.0')
   expect(output.date).toBeUndefined()
-  expect(output.text).toContain(`### Added`)
-  expect(output.text).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
-  expect(output.text).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+  expect(output.status).toEqual('released')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+})
+
+test('get readable data from text entry | build number with plus sign', () => {
+  const input = `
+    ## [2.0.0+build.1848] - SomeWords instead of a Date
+    ${entryDescription}
+  `
+  const output = parseEntry(input)
+
+  expect(output.id).toEqual('2.0.0+build.1848')
+  expect(output.date).toBeUndefined()
+  expect(output.status).toEqual('released')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+})
+
+test('get readable data from text entry | yanked release', () => {
+  const input = `
+    ## [2.0.0] - 2019-02-10 [YANKED]
+    ${entryDescription}
+  `
+  const output = parseEntry(input)
+
+  expect(output.id).toEqual('2.0.0')
+  expect(output.date).toEqual('2019-02-10')
+  expect(output.status).toEqual('yanked')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
+})
+
+test('get readable data from text entry | yanked release with no date', () => {
+  const input = `
+    ## [2.0.0] [YANKED]
+    ${entryDescription}
+  `
+  const output = parseEntry(input)
+
+  expect(output.id).toEqual('2.0.0')
+  expect(output.date).toBeUndefined()
+  expect(output.status).toEqual('yanked')
+  expect(output.changes).toContain(`### Added`)
+  expect(output.changes).toContain(`ThemeProvider doesn't loads the font anymore. We created a more generic component (UIKitInitializer) that'll do it.`)
+  expect(output.changes).toContain(`**SECURITY** The list components don't use the nth-child CSS attributes in favor of nth-of-type.`)
 })