+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- **[BREAKING CHANGE]** If given a specific target version, action will now generate an error response if that version is not found in the changelog.
 
 ## [1.3.1] - 2020-07-08
 ### Fixed