+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2020-08-30
 ### Added
 - New output properties:
   - The `version` number of the returned entry
@@ -58,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - CHANGELOG can be parsed by the github action
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/mindsers/changelog-reader-action/compare/v1.3.1...v2.0.0
 [1.3.1]: https://github.com/mindsers/changelog-reader-action/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/mindsers/changelog-reader-action/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/mindsers/changelog-reader-action/compare/v1.1.0...v1.2.0